video_core: Move surface declarations out of gl_rasterizer_cache
